Move evaluate_and_log_point to optimize_result
